auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Local authentication service for manager and coach roles"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.29",
  "fastapi>=0.115",
  "pydantic[email]>=2.8",
  "pyjwt>=2.9",
  "python-dotenv>=1.0",
  "uvicorn>=0.30",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "database", "security"]
